Updates for various deprecations with clang/10 and c++20

+++ core/unit_test/TestTeamVector.hpp
@@ -109,7 +109,7 @@ struct functor_team_for {
           value += values(i);
         }
 
-        if (test != value) {
+        if (static_cast<double>(test) != static_cast<double>(value)) {
           printf("FAILED team_parallel_for %i %i %f %f\n", team.league_rank(),
                  team.team_rank(), static_cast<double>(test),
                  static_cast<double>(value));
@@ -166,7 +166,7 @@ struct functor_team_reduce {
         test += i - team.league_rank() + team.league_size() + team.team_size();
       }
 
-      if (test != value) {
+      if (static_cast<double>(test) != static_cast<double>(value)) {
         if (team.league_rank() == 0) {
           printf("FAILED team_parallel_reduce %i %i %lf %lf %lu\n",
                  team.league_rank(), team.team_rank(),
@@ -176,7 +176,7 @@ struct functor_team_reduce {
 
         flag() = 1;
       }
-      if (test != shared_value(0)) {
+      if (static_cast<double>(test) != static_cast<double>(shared_value(0))) {
         if (team.league_rank() == 0) {
           printf(
               "FAILED team_parallel_reduce with shared result %i %i %lf %lf "
@@ -238,14 +238,14 @@ struct functor_team_reduce_reducer {
         test += i - team.league_rank() + team.league_size() + team.team_size();
       }
 
-      if (test != value) {
+      if (static_cast<double>(test) != static_cast<double>(value)) {
         printf("FAILED team_vector_parallel_reduce_reducer %i %i %lf %lf\n",
                team.league_rank(), team.team_rank(), static_cast<double>(test),
                static_cast<double>(value));
 
         flag() = 1;
       }
-      if (test != shared_value(0)) {
+      if (static_cast<double>(test) != static_cast<double>(shared_value(0))) {
         printf(
             "FAILED team_vector_parallel_reduce_reducer shared value %i %i %lf "
             "%lf\n",
@@ -316,8 +316,8 @@ struct functor_team_vector_for {
           value += values(i);
         }
 
-        if (test != value) {
-          printf("FAILED team_vector_parallel_for %i %i %f %f\n",
+        if (static_cast<double>(test) != static_cast<double>(value)) {
+          printf("FAILED team_vector_parallel_for %i %i %lf %lf\n",
                  team.league_rank(), team.team_rank(),
                  static_cast<double>(test), static_cast<double>(value));
 
@@ -365,9 +365,9 @@ struct functor_team_vector_reduce {
         test += i - team.league_rank() + team.league_size() + team.team_size();
       }
 
-      if (test != value) {
+      if (static_cast<double>(test) != static_cast<double>(value)) {
         if (team.league_rank() == 0) {
-          printf("FAILED team_vector_parallel_reduce %i %i %f %f %lu\n",
+          printf("FAILED team_vector_parallel_reduce %i %i %lf %lf %lu\n",
                  team.league_rank(), team.team_rank(),
                  static_cast<double>(test), static_cast<double>(value),
                  static_cast<unsigned long>(sizeof(Scalar)));
@@ -417,8 +417,8 @@ struct functor_team_vector_reduce_reducer {
         test += i - team.league_rank() + team.league_size() + team.team_size();
       }
 
-      if (test != value) {
-        printf("FAILED team_vector_parallel_reduce_reducer %i %i %f %f\n",
+      if (static_cast<double>(test) != static_cast<double>(value)) {
+        printf("FAILED team_vector_parallel_reduce_reducer %i %i %lf %lf\n",
                team.league_rank(), team.team_rank(), static_cast<double>(test),
                static_cast<double>(value));
 
@@ -465,8 +465,9 @@ struct functor_vec_single {
 
     if (value2 != (value * Scalar(nEnd - nStart))) {
       KOKKOS_IMPL_DO_NOT_USE_PRINTF(
-          "FAILED vector_single broadcast %i %i %f %f\n", team.league_rank(),
-          team.team_rank(), (double)value2, (double)value);
+          "FAILED vector_single broadcast %i %i %lf %lf\n", team.league_rank(),
+          team.team_rank(), static_cast<double>(value2),
+          static_cast<double>(value));
 
       flag() = 1;
     }
@@ -516,7 +517,7 @@ struct functor_vec_for {
           value += values(13 * team.team_rank() + i);
         }
 
-        if (test != value) {
+        if (static_cast<double>(test) != static_cast<double>(value)) {
           printf("FAILED vector_par_for %i %i %f %f\n", team.league_rank(),
                  team.team_rank(), static_cast<double>(test),
                  static_cast<double>(value));
@@ -552,9 +553,11 @@ struct functor_vec_red {
 
       for (int i = 0; i < 13; i++) test += i;
 
-      if (test != value) {
-        printf("FAILED vector_par_reduce %i %i %f %f\n", team.league_rank(),
-               team.team_rank(), (double)test, (double)value);
+      // use of overloaded operator '!=' is ambiguous
+      if (static_cast<double>(test) != static_cast<double>(value)) {
+        printf("FAILED vector_par_reduce %i %i %lf %lf\n", team.league_rank(),
+               team.team_rank(), static_cast<double>(test),
+               static_cast<double>(value));
 
         flag() = 1;
       }
@@ -590,10 +593,10 @@ struct functor_vec_red_reducer {
 
       for (int i = 0; i < 13; i++) test *= (i % 5 + 1);
 
-      if (test != value) {
-        printf("FAILED vector_par_reduce_reducer %i %i %f %f\n",
-               team.league_rank(), team.team_rank(), (double)test,
-               (double)value);
+      if (static_cast<double>(test) != static_cast<double>(value)) {
+        printf("FAILED vector_par_reduce_reducer %i %i %lf %lf\n",
+               team.league_rank(), team.team_rank(), static_cast<double>(test),
+               static_cast<double>(value));
 
         flag() = 1;
       }
@@ -612,23 +615,24 @@ struct functor_vec_scan {
 
   KOKKOS_INLINE_FUNCTION
   void operator()(typename policy_type::member_type team) const {
-    Kokkos::parallel_scan(Kokkos::ThreadVectorRange(team, 13),
-                          [&](int i, Scalar &val, bool final) {
-                            val += i;
+    Kokkos::parallel_scan(
+        Kokkos::ThreadVectorRange(team, 13),
+        [&](int i, Scalar &val, bool final) {
+          val += i;
 
-                            if (final) {
-                              Scalar test = 0;
-                              for (int k = 0; k <= i; k++) test += k;
+          if (final) {
+            Scalar test = 0;
+            for (int k = 0; k <= i; k++) test += k;
 
-                              if (test != val) {
-                                printf("FAILED vector_par_scan %i %i %f %f\n",
-                                       team.league_rank(), team.team_rank(),
-                                       (double)test, (double)val);
+            if (static_cast<double>(test) != static_cast<double>(val)) {
+              printf("FAILED vector_par_scan %i %i %lf %lf\n",
+                     team.league_rank(), team.team_rank(),
+                     static_cast<double>(test), static_cast<double>(val));
 
-                                flag() = 1;
-                              }
-                            }
-                          });
+              flag() = 1;
+            }
+          }
+        });
   }
 };
